Add files via upload
Day5. This is a 93,526 data rows for the New York City Police Department crimes unit that we are cleaning for the data visualization project during this #21DaystoData by Avery Smith. The data is pretty dense with not a lot of aggregated columns. The assignment was to look at the columns that needed to be cleaned, added some new columns to the data, create the Day of the week column, Hour of crime column, time of crime, I went ahead to create a Boolean for the RM_ATPT_CPTD_CD Column which gave a YES or NO response if the crime was completed or not, There was a lot of blank space in the Park_NM  column but cannot filter it out so I wrote a condition to put null in the blank spaces and still keep those with park names in order to be able to work with it.